update: utils created and it maintans shared methods now (#69)
